remove ts
