+++ gatsby/gatsby-node.js
@@ -1,0 +1,11 @@
+export const createSchemaCustomization = ({ actions }) => {
+  actions.createTypes(`
+    type MarkdownRemark implements Node @dontInfer {
+      frontmatter: Frontmatter
+      html: String!
+    }
+    type Frontmatter {
+      title: String!
+    }
+  `);
+};
