--- gatsby/gatsby-node.ts
@@ -1,14 +1,0 @@
-import type { GatsbyNode } from "gatsby";
-
-export const createSchemaCustomization: GatsbyNode["createSchemaCustomization"] =
-  ({ actions }) => {
-    actions.createTypes(`
-    type MarkdownRemark implements Node @dontInfer {
-      frontmatter: Frontmatter
-      html: String!
-    }
-    type Frontmatter {
-      title: String!
-    }
-  `);
-  };
